refactor(components): split group lifecycle
